[REFACTOR] moved select package into media package
